Refactor sparse utils and eliminate duplication
