+++ cellbender/remove_background/sparse_utils.py
@@ -4,7 +4,7 @@ import torch
 import numpy as np
 import scipy.sparse as sp
 
-from typing import Optional, Tuple
+from typing import Optional, Tuple, Iterable
 
 
 @torch.no_grad()
@@ -55,16 +55,47 @@ def todense_fill(coo: sp.coo_matrix, fill_value: float) -> np.ndarray:
     return out
 
 
-def zero_out_csr_rows(csr: sp.csr_matrix,
-                      row_logic: np.ndarray) -> sp.csr_matrix:
-    """Given a sparse matrix, set specified rows to zero.
+def csr_set_rows_to_zero(csr: sp.csr_matrix,
+                         row_inds: Iterable[int]) -> sp.csr_matrix:
+    """Set all nonzero elements in rows "row_inds" to zero.
+    Happens in-place, although output is returned as well.
+
+    https://stackoverflow.com/questions/12129948/scipy-sparse-set-row-to-zeros
     """
-    row_inds = set(np.where(row_logic)[0])
-    coo = csr.copy().tocoo()
 
-    # Zero out values that are in the specified rows.
-    logic = np.array([i in row_inds for i in coo.row])
-    coo.data[logic] = 0
-    coo.eliminate_zeros()
+    if not isinstance(csr, sp.csr_matrix):
+        try:
+            csr = csr.tocsr()
+        except Exception:
+            raise ValueError('Matrix given must be of CSR format.')
+    for row in row_inds:
+        csr.data[csr.indptr[row]:csr.indptr[row + 1]] = 0
+    csr.eliminate_zeros()
+    return csr
 
-    return coo.tocsr()
+
+def overwrite_matrix_with_columns_from_another(mat1: sp.csc_matrix,
+                                               mat2: sp.csc_matrix,
+                                               column_inds: np.ndarray) -> sp.csc_matrix:
+    """Given two sparse matrices of the same shape, replace columns that are not
+    in `column_inds` in `mat1` with the entries from `mat2`.
+    """
+    column_inds = set(column_inds)
+
+    mat1 = mat1.copy().tocsr()
+    mat2 = mat2.copy().tocsr()  # failure to copy could overwrite actual count data
+
+    # Zero out values in mat2 that are in the specified columns.
+    inds = np.where([i in column_inds for i in mat2.indices])[0]
+    mat2.data[inds] = 0
+    mat2.eliminate_zeros()
+
+    # Zero out values in mat1 that are not in the specified columns.
+    inds = np.where([i not in column_inds for i in mat1.indices])[0]
+    mat1.data[inds] = 0
+    mat1.eliminate_zeros()
+
+    # Put in the new values by addition.
+    output = mat1 + mat2
+
+    return output.tocsc()
